feat(ci): add pre-commit hooks for format and lint staging
Added `husky` and `lint-staged` to the project to enable running biome format and lintcommands on all JavaScript, JSX, TypeScript, and TSX files, as well as updating the`bun.lockb` on changes to `bun.lockb`. This ensures consistent code formatting and
linting before committing, contributing to easier code reviews and less time spent on
formatting during the development process.

Signed-off-by: yebei199 <scauyb@qq.com>